// Qanat.Web/src/app/shared/generated/model/meter-reading-monthly-interpolation-simple-dto.ts
/**
 * Qanat.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class MeterReadingMonthlyInterpolationSimpleDto { 
    MeterReadingMonthlyInterpolationID?: number;
    GeographyID?: number;
    WellID?: number;
    MeterID?: number;
    MeterReadingUnitTypeID?: number;
    Date?: string;
    InterpolatedVolume?: number;
    InterpolatedVolumeInAcreFeet?: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface MeterReadingMonthlyInterpolationSimpleDtoForm { 
    MeterReadingMonthlyInterpolationID?: FormControl<number>;
    GeographyID?: FormControl<number>;
    WellID?: FormControl<number>;
    MeterID?: FormControl<number>;
    MeterReadingUnitTypeID?: FormControl<number>;
    Date?: FormControl<string>;
    InterpolatedVolume?: FormControl<number>;
    InterpolatedVolumeInAcreFeet?: FormControl<number>;
}

export class MeterReadingMonthlyInterpolationSimpleDtoFormControls { 
    public static MeterReadingMonthlyInterpolationID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static GeographyID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static WellID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static MeterID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static MeterReadingUnitTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Date = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static InterpolatedVolume = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static InterpolatedVolumeInAcreFeet = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}