// Qanat.Web/src/app/shared/generated/model/well-file-resource-dto.ts
/**
 * Qanat.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FileResourceSimpleDto } from './file-resource-simple-dto';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class WellFileResourceDto { 
    WellFileResourceID?: number;
    WellID?: number;
    FileResourceID?: number;
    FileDescription?: string | null;
    FileResource?: FileResourceSimpleDto;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface WellFileResourceDtoForm { 
    WellFileResourceID?: FormControl<number>;
    WellID?: FormControl<number>;
    FileResourceID?: FormControl<number>;
    FileDescription?: FormControl<string>;
    FileResource?: FormControl<FileResourceSimpleDto>;
}

export class WellFileResourceDtoFormControls { 
    public static WellFileResourceID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static WellID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static FileResourceID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static FileDescription = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static FileResource = (value: FormControlState<FileResourceSimpleDto> | FileResourceSimpleDto = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<FileResourceSimpleDto>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}